EPMRPP-110765 || refactor imports to use base package structure
