+++ build.gradle
@@ -29,6 +29,7 @@ apply from: scriptsUrl + '/signing.gradle'
 repositories {
     mavenLocal()
     mavenCentral { url "https://repo1.maven.org/maven2" }
+    maven { url "https://build.shibboleth.net/nexus/content/repositories/releases/" }
     if (!releaseMode) {
         maven { url 'https://jitpack.io' }
     }
@@ -41,8 +42,8 @@ dependencies {
         implementation 'com.epam.reportportal:service-api'
         annotationProcessor 'com.epam.reportportal:service-api'
     } else {
-        implementation 'com.github.reportportal:service-api:c200fb8'
-        annotationProcessor 'com.github.reportportal:service-api:c200fb8'
+        implementation 'com.github.reportportal:service-api:5f87c3c'
+        annotationProcessor 'com.github.reportportal:service-api:5f87c3c'
     }
     implementation 'com.fasterxml.jackson.core:jackson-databind'
     implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310'
