+++ src/main/java/com/epam/reportportal/extension/importing/command/XUnitImportCommand.java
@@ -3,9 +3,9 @@ package com.epam.reportportal.extension.importing.command;
 import static com.epam.reportportal.extension.importing.service.FileExtensionConstant.XML_EXTENSION;
 import static com.epam.reportportal.extension.importing.service.FileExtensionConstant.ZIP_EXTENSION;
 import static com.epam.reportportal.extension.util.CommandParamUtils.ENTITY_PARAM;
-import static com.epam.reportportal.infrastructure.rules.commons.validation.BusinessRule.expect;
-import static com.epam.reportportal.infrastructure.rules.exception.ErrorType.BAD_REQUEST_ERROR;
-import static com.epam.reportportal.infrastructure.rules.exception.ErrorType.INCORRECT_REQUEST;
+import static com.epam.reportportal.base.infrastructure.rules.commons.validation.BusinessRule.expect;
+import static com.epam.reportportal.base.infrastructure.rules.exception.ErrorType.BAD_REQUEST_ERROR;
+import static com.epam.reportportal.base.infrastructure.rules.exception.ErrorType.INCORRECT_REQUEST;
 import static org.apache.commons.io.FileUtils.ONE_MB;
 
 import com.epam.reportportal.extension.CommonPluginCommand;
@@ -13,9 +13,9 @@ import com.epam.reportportal.extension.importing.model.LaunchImportRQ;
 import com.epam.reportportal.extension.importing.service.ImportStrategy;
 import com.epam.reportportal.extension.importing.service.ImportStrategyFactory;
 import com.epam.reportportal.extension.util.RequestEntityConverter;
-import com.epam.reportportal.infrastructure.persistence.dao.LaunchRepository;
-import com.epam.reportportal.infrastructure.rules.exception.ReportPortalException;
-import com.epam.reportportal.reporting.StartLaunchRS;
+import com.epam.reportportal.base.infrastructure.persistence.dao.LaunchRepository;
+import com.epam.reportportal.base.infrastructure.rules.exception.ReportPortalException;
+import com.epam.reportportal.base.reporting.StartLaunchRS;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
