+++ src/main/java/com/epam/reportportal/extension/importing/model/ItemInfo.java
@@ -1,6 +1,6 @@
 package com.epam.reportportal.extension.importing.model;
 
-import com.epam.reportportal.reporting.ItemAttributesRQ;
+import com.epam.reportportal.base.reporting.ItemAttributesRQ;
 import java.time.Instant;
 import java.util.Set;
