+++ src/main/java/com/epam/reportportal/extension/importing/model/LaunchImportCompletionRS.java
@@ -16,7 +16,7 @@
 
 package com.epam.reportportal.extension.importing.model;
 
-import com.epam.reportportal.reporting.OperationCompletionRS;
+import com.epam.reportportal.base.reporting.OperationCompletionRS;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 /**
