+++ src/main/java/com/epam/reportportal/extension/importing/service/AbstractImportStrategy.java
@@ -19,17 +19,17 @@ package com.epam.reportportal.extension.importing.service;
 import static java.util.Optional.ofNullable;
 
 import com.epam.reportportal.extension.importing.model.LaunchImportRQ;
-import com.epam.reportportal.infrastructure.events.FinishLaunchRqEvent;
-import com.epam.reportportal.infrastructure.events.StartLaunchRqEvent;
-import com.epam.reportportal.infrastructure.persistence.dao.LaunchRepository;
-import com.epam.reportportal.infrastructure.persistence.entity.enums.StatusEnum;
-import com.epam.reportportal.infrastructure.persistence.entity.launch.Launch;
-import com.epam.reportportal.infrastructure.rules.exception.ErrorType;
-import com.epam.reportportal.infrastructure.rules.exception.ReportPortalException;
-import com.epam.reportportal.reporting.FinishExecutionRQ;
-import com.epam.reportportal.reporting.ItemAttributesRQ;
-import com.epam.reportportal.reporting.Mode;
-import com.epam.reportportal.reporting.StartLaunchRQ;
+import com.epam.reportportal.base.infrastructure.events.FinishLaunchRqEvent;
+import com.epam.reportportal.base.infrastructure.events.StartLaunchRqEvent;
+import com.epam.reportportal.base.infrastructure.persistence.dao.LaunchRepository;
+import com.epam.reportportal.base.infrastructure.persistence.entity.enums.StatusEnum;
+import com.epam.reportportal.base.infrastructure.persistence.entity.launch.Launch;
+import com.epam.reportportal.base.infrastructure.rules.exception.ErrorType;
+import com.epam.reportportal.base.infrastructure.rules.exception.ReportPortalException;
+import com.epam.reportportal.base.reporting.FinishExecutionRQ;
+import com.epam.reportportal.base.reporting.ItemAttributesRQ;
+import com.epam.reportportal.base.reporting.Mode;
+import com.epam.reportportal.base.reporting.StartLaunchRQ;
 import java.time.Instant;
 import java.util.Collections;
 import java.util.HashSet;
