+++ src/main/java/com/epam/reportportal/extension/importing/service/XmlImportStrategy.java
@@ -19,9 +19,9 @@ package com.epam.reportportal.extension.importing.service;
 import static com.epam.reportportal.extension.importing.service.FileExtensionConstant.XML_EXTENSION;
 
 import com.epam.reportportal.extension.importing.model.LaunchImportRQ;
-import com.epam.reportportal.infrastructure.persistence.dao.LaunchRepository;
-import com.epam.reportportal.infrastructure.rules.exception.ErrorType;
-import com.epam.reportportal.infrastructure.rules.exception.ReportPortalException;
+import com.epam.reportportal.base.infrastructure.persistence.dao.LaunchRepository;
+import com.epam.reportportal.base.infrastructure.rules.exception.ErrorType;
+import com.epam.reportportal.base.infrastructure.rules.exception.ReportPortalException;
 import java.io.InputStream;
 import java.util.Optional;
 import org.springframework.context.ApplicationEventPublisher;
