+++ src/main/java/com/epam/reportportal/extension/importing/service/XunitImportHandler.java
@@ -24,24 +24,24 @@ import static com.epam.reportportal.extension.importing.service.XunitReportTag.T
 import static com.epam.reportportal.extension.importing.service.XunitReportTag.TIMESTAMP;
 import static com.epam.reportportal.extension.importing.service.XunitReportTag.fromString;
 import static com.epam.reportportal.extension.importing.utils.DateUtils.toMillis;
-import static com.epam.reportportal.infrastructure.persistence.entity.enums.TestItemIssueGroup.NOT_ISSUE_FLAG;
+import static com.epam.reportportal.base.infrastructure.persistence.entity.enums.TestItemIssueGroup.NOT_ISSUE_FLAG;
 import static java.util.Optional.ofNullable;
 
 import com.epam.reportportal.extension.importing.model.ItemInfo;
-import com.epam.reportportal.infrastructure.events.FinishItemRqEvent;
-import com.epam.reportportal.infrastructure.events.SaveLogRqEvent;
-import com.epam.reportportal.infrastructure.events.StartChildItemRqEvent;
-import com.epam.reportportal.infrastructure.events.StartRootItemRqEvent;
-import com.epam.reportportal.infrastructure.persistence.entity.enums.LogLevel;
-import com.epam.reportportal.infrastructure.persistence.entity.enums.StatusEnum;
-import com.epam.reportportal.infrastructure.persistence.entity.enums.TestItemTypeEnum;
-import com.epam.reportportal.infrastructure.rules.exception.ErrorType;
-import com.epam.reportportal.infrastructure.rules.exception.ReportPortalException;
-import com.epam.reportportal.reporting.FinishTestItemRQ;
-import com.epam.reportportal.reporting.Issue;
-import com.epam.reportportal.reporting.ItemAttributesRQ;
-import com.epam.reportportal.reporting.SaveLogRQ;
-import com.epam.reportportal.reporting.StartTestItemRQ;
+import com.epam.reportportal.base.infrastructure.events.FinishItemRqEvent;
+import com.epam.reportportal.base.infrastructure.events.SaveLogRqEvent;
+import com.epam.reportportal.base.infrastructure.events.StartChildItemRqEvent;
+import com.epam.reportportal.base.infrastructure.events.StartRootItemRqEvent;
+import com.epam.reportportal.base.infrastructure.persistence.entity.enums.LogLevel;
+import com.epam.reportportal.base.infrastructure.persistence.entity.enums.StatusEnum;
+import com.epam.reportportal.base.infrastructure.persistence.entity.enums.TestItemTypeEnum;
+import com.epam.reportportal.base.infrastructure.rules.exception.ErrorType;
+import com.epam.reportportal.base.infrastructure.rules.exception.ReportPortalException;
+import com.epam.reportportal.base.reporting.FinishTestItemRQ;
+import com.epam.reportportal.base.reporting.Issue;
+import com.epam.reportportal.base.reporting.ItemAttributesRQ;
+import com.epam.reportportal.base.reporting.SaveLogRQ;
+import com.epam.reportportal.base.reporting.StartTestItemRQ;
 import com.google.common.base.Strings;
 import com.google.common.collect.Lists;
 import java.time.Instant;
